+++ src/devices/lanesto.ts
@@ -3,7 +3,7 @@ import {Definition} from '../lib/types';
 
 const definitions: Definition[] = [
     {
-        zigbeeModel: ['FB56-ZBW14LF1.4', 'RH0039'],
+        zigbeeModel: ['FB56-ZBW14LF1.4', 'RH0039', 'AG0002'],
         model: '322054',
         vendor: 'Lanesto',
         description: 'Dimmable led driver',