+++ src/lib/ikea.ts
@@ -86,7 +86,7 @@ const bulbOnEvent: OnEvent = async (type, data, device, options, state: KeyValue
     }
 };
 
-export function ikeaLight(args?: Omit<LightArgs, 'colorTemp'> & {colorTemp?: true | {range: Range; viaColor: true}}) {
+export function ikeaLight(args?: Omit<LightArgs, 'colorTemp'> & {colorTemp?: true | {range: Range; viaColor: true}; noOffTransition?: true}) {
     const colorTemp: {range: Range} = args?.colorTemp ? (args.colorTemp === true ? {range: [250, 454]} : args.colorTemp) : undefined;
     const result = lightDontUse({...args, colorTemp});
     result.ota = ikea;
@@ -96,6 +96,9 @@ export function ikeaLight(args?: Omit<LightArgs, 'colorTemp'> & {colorTemp?: tru
     }
     if (args?.colorTemp || args?.color) {
         result.exposes.push(presets.light_color_options());
+    }
+    if (args?.noOffTransition) {
+        result.meta = {...result.meta, noOffTransition: true};
     }
     return result;
 }